Move files and walrus operator
