+++ src/motile_tracker/data_views/views/table/custom_table_widget.py
