+++ tests/data_views/views/table/test_table_widget.py
@@ -2,7 +2,7 @@ import pandas as pd
 import pytest
 from qtpy.QtWidgets import QApplication
 
-from motile_tracker.data_views.views.tree_view.custom_table_widget import (
+from motile_tracker.data_views.views.table.custom_table_widget import (
     ColoredTableWidget,
 )
 from motile_tracker.data_views.views_coordinator.tracks_viewer import TracksViewer
